requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements3.txt requirements3.in
#
click==6.7                # via pip-tools
first==2.0.1              # via pip-tools
invoke==0.12.2
jinja2==2.9.4
MarkupSafe==0.23          # via jinja2
nose==1.3.7
olefile==0.44             # via pillow
pillow==4.0.0
pip-tools==1.8.0
pyyaml==3.12
six==1.10.0               # via pip-tools
